24.04.2020, 13:48

+++ src/sample/Tools.java
@@ -7,7 +7,7 @@ public class Tools {
     private long price;
     private int level;
     private boolean status;
-    private long moneyPerSecond;
+    private double moneyPerSecond;
     private int multiplier;
     private long pricePerLevel;
     ToolsJDBCDoa Tools = new ToolsJDBCDoa();
@@ -16,7 +16,7 @@ public class Tools {
     /*------------------------------------------------
                         Constructor
     ------------------------------------------------*/
-    public Tools(String name, long price, int level, boolean status, long moneyPerSecond, int multiplier, long pricePerLevel) {
+    public Tools(String name, long price, int level, boolean status, double moneyPerSecond, int multiplier, long pricePerLevel) {
         this.name = name;
         this.price = price;
         this.level = level;
@@ -51,7 +51,7 @@ public class Tools {
         return multiplier;
     }
 
-    public long getMoneyPerSecond() {
+    public double getMoneyPerSecond() {
         return moneyPerSecond;
     }
 
@@ -77,7 +77,7 @@ public class Tools {
         this.multiplier = multiplier;
     }
 
-    public void setMoneyPerSecond(long moneyPerSecond) { this.moneyPerSecond = moneyPerSecond; }
+    public void setMoneyPerSecond(double moneyPerSecond) { this.moneyPerSecond = moneyPerSecond; }
 
     public void setStatus(boolean status) {
         this.status = status;
